+++ packages/cli/src/commands/org/config/preview.ts
@@ -12,6 +12,8 @@ import {Project} from '../../../lib/project/project';
 import {SnapshotFactory} from '../../../lib/snapshot/snapshotFactory';
 import {platformUrl} from '../../../lib/platform/environment';
 import {Snapshot} from '../../../lib/snapshot/snapshot';
+import {red, green} from 'chalk';
+import {normalize} from 'path';
 
 export interface CustomFile extends ReadStream {
   type?: string;
@@ -35,6 +37,12 @@ export default class Preview extends Command {
       default: cwd(),
       required: false,
     }),
+    showResourcesToDelete: flags.boolean({
+      char: 'd',
+      description: 'Whether or not to show resources to delete',
+      default: false,
+      required: false,
+    }),
   };
 
   public static hidden = true;
@@ -42,7 +50,7 @@ export default class Preview extends Command {
   @Preconditions(IsAuthenticated())
   public async run() {
     const {flags} = this.parse(Preview);
-    const project = new Project(flags.projectPath);
+    const project = new Project(normalize(flags.projectPath));
     const pathToZip = await project.compressResources();
     const targetOrg = await this.getTargetOrg();
 
@@ -52,18 +60,21 @@ export default class Preview extends Command {
 
     cli.action.start('Validating snapshot');
 
-    const {isValid} = await snapshot.validate();
+    const {isValid} = await snapshot.validate(flags.showResourcesToDelete);
 
     if (!isValid) {
-      this.handleInvalidSnapshot(snapshot);
-    } else {
-      await snapshot.preview();
+      await this.handleInvalidSnapshot(snapshot);
+    }
+
+    cli.action.stop(isValid ? green('✔') : red.bold('!'));
+
+    await snapshot.preview();
+
+    if (isValid) {
       await snapshot.delete();
     }
 
     project.deleteTemporaryZipFile();
-
-    cli.action.stop();
   }
 
   public async getTargetOrg() {
@@ -82,20 +93,20 @@ export default class Preview extends Command {
     const report = snapshot.latestReport;
 
     if (snapshot.requiresSynchronization()) {
-      cli.action.start('Synchronization');
-
       const synchronizationPlanUrl = await this.getSynchronizationPage(
         snapshot
       );
       this.warn(
-        dedent`Some conflicts were detected while comparing changes between the snapshot and the target organization.
+        dedent`
+        Some conflicts were detected while comparing changes between the snapshot and the target organization.
         Click on the URL below to synchronize your snapshot with your organization before running the command again.
-        ${synchronizationPlanUrl}`
+        ${synchronizationPlanUrl}
+        `
       );
       return;
     }
 
-    const snapshotUrl = this.getSnapshotPage(snapshot);
+    const snapshotUrl = await this.getSnapshotPage(snapshot);
 
     this.error(
       dedent`Invalid snapshot - ${report.resultCode}.